show render output
